requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-Migrate
python-dotenv
psycopg2-binary
gunicorn
Flask-WTF
python-dateutil